Move builders related classes to bitcoin package and create RedeemScriptCreationException class to avoid circular dependency with peg package
