+++ rskj-core/src/main/java/co/rsk/peg/ErpFederation.java
@@ -7,12 +7,15 @@ import co.rsk.bitcoinj.script.RedeemScriptParserFactory;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
 import co.rsk.bitcoinj.script.ScriptChunk;
+import co.rsk.peg.bitcoin.ErpRedeemScriptBuilder;
+import co.rsk.peg.bitcoin.RedeemScriptCreationException;
 import co.rsk.peg.utils.EcKeyUtils;
 import java.time.Instant;
 import java.util.Collections;
 import java.util.List;
 
 import static co.rsk.peg.ErpFederationCreationException.Reason.NULL_OR_EMPTY_EMERGENCY_KEYS;
+import static co.rsk.peg.ErpFederationCreationException.Reason.REDEEM_SCRIPT_CREATION_FAILED;
 
 public class ErpFederation extends Federation {
     private final List<BtcECKey> erpPubKeys;
@@ -51,7 +54,6 @@ public class ErpFederation extends Federation {
         return Collections.unmodifiableList(erpPubKeys);
     }
 
-
     public int getNumberOfEmergencySignaturesRequired() {
         return erpPubKeys.size() / 2 + 1;
     }
@@ -71,6 +73,7 @@ public class ErpFederation extends Federation {
     @Override
     public Script getRedeemScript() {
         if (redeemScript == null) {
+            try {
                 redeemScript = erpRedeemScriptBuilder.createRedeemScriptFromKeys(
                     getBtcPublicKeys(),
                     getNumberOfSignaturesRequired(),
@@ -78,6 +81,9 @@ public class ErpFederation extends Federation {
                     getNumberOfEmergencySignaturesRequired(),
                     activationDelay
                 );
+            } catch (RedeemScriptCreationException e) {
+                throw new ErpFederationCreationException(e.getMessage(), e, REDEEM_SCRIPT_CREATION_FAILED);
+            }
         }
         return redeemScript;
     }
