+++ rskj-core/src/main/java/co/rsk/peg/StandardMultisigFederation.java
@@ -21,6 +21,7 @@ package co.rsk.peg;
 import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
+import co.rsk.peg.bitcoin.ScriptValidations;
 
 import java.time.Instant;
 import java.util.List;
