+++ rskj-core/src/main/java/co/rsk/peg/bitcoin/NonStandardErpRedeemScriptBuilderWithCsvUnsignedBE.java
@@ -1,4 +1,4 @@
-package co.rsk.peg;
+package co.rsk.peg.bitcoin;
 
 import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.Utils;
@@ -10,8 +10,11 @@ import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
-public class P2shErpRedeemScriptBuilder implements ErpRedeemScriptBuilder{
-    private static final Logger logger = LoggerFactory.getLogger(P2shErpRedeemScriptBuilder.class);
+import static co.rsk.peg.bitcoin.ErpRedeemScriptBuilderUtils.removeOpCheckMultisig;
+
+public class NonStandardErpRedeemScriptBuilderWithCsvUnsignedBE implements ErpRedeemScriptBuilder {
+    private static final Logger logger = LoggerFactory.getLogger(NonStandardErpRedeemScriptBuilderWithCsvUnsignedBE.class);
+    private static final int CSV_BYTES_NEEDED_LENGTH = 2;
 
     @Override
     public Script createRedeemScriptFromKeys(List<BtcECKey> defaultPublicKeys,
@@ -25,30 +28,32 @@ public class P2shErpRedeemScriptBuilder implements ErpRedeemScriptBuilder{
 
         ErpRedeemScriptBuilderUtils.validateRedeemScriptValues(defaultRedeemScript, emergencyRedeemScript, csvValue);
 
-        byte[] serializedCsvValue = Utils.signedLongToByteArrayLE(csvValue);
+        byte[] serializedCsvValue = Utils.unsignedLongToByteArrayBE(csvValue, CSV_BYTES_NEEDED_LENGTH);
         logger.debug("[createRedeemScriptFromKeys] Creating the redeem script from the scripts");
         Script redeemScript = createRedeemScriptFromScripts(defaultRedeemScript, emergencyRedeemScript, serializedCsvValue);
-
+        
         logger.debug("[createRedeemScriptFromKeys] Validating redeem script size");
         ScriptValidations.validateScriptSize(redeemScript);
-
+        
         return redeemScript;
+
     }
+
     private Script createRedeemScriptFromScripts(Script defaultRedeemScript,
-                                     Script emergencyRedeemScript,
-                                     byte[] serializedCsvValue) {
+                                                        Script emergencyRedeemScript,
+                                                        byte[] serializedCsvValue) {
 
         ScriptBuilder scriptBuilder = new ScriptBuilder();
-
         return scriptBuilder
             .op(ScriptOpCodes.OP_NOTIF)
-            .addChunks(defaultRedeemScript.getChunks())
+            .addChunks(removeOpCheckMultisig(defaultRedeemScript))
             .op(ScriptOpCodes.OP_ELSE)
             .data(serializedCsvValue)
             .op(ScriptOpCodes.OP_CHECKSEQUENCEVERIFY)
             .op(ScriptOpCodes.OP_DROP)
-            .addChunks(emergencyRedeemScript.getChunks())
+            .addChunks(removeOpCheckMultisig(emergencyRedeemScript))
             .op(ScriptOpCodes.OP_ENDIF)
+            .op(ScriptOpCodes.OP_CHECKMULTISIG)
             .build();
     }
 }
