+++ rskj-core/src/main/java/co/rsk/peg/bitcoin/ScriptValidations.java
@@ -1,13 +1,13 @@
-package co.rsk.peg;
+package co.rsk.peg.bitcoin;
 
 import co.rsk.bitcoinj.script.Script;
 
-import static co.rsk.peg.ScriptCreationException.Reason.ABOVE_MAX_SCRIPT_ELEMENT_SIZE;
+import static co.rsk.peg.bitcoin.ScriptCreationException.Reason.ABOVE_MAX_SCRIPT_ELEMENT_SIZE;
 
 public class ScriptValidations {
     private ScriptValidations() {
     }
-    public static void validateScriptSize(Script script) throws ErpFederationCreationException {
+    public static void validateScriptSize(Script script) throws ScriptCreationException {
         // Check if the size of the script does not exceed the maximum size allowed
         int bytesFromScript = script.getProgram().length;
         if (bytesFromScript > Script.MAX_SCRIPT_ELEMENT_SIZE) {
