+++ rskj-core/src/test/java/co/rsk/peg/BridgeStorageProviderFederationTests.java
@@ -14,6 +14,9 @@ import java.io.IOException;
 import java.time.Instant;
 import java.util.List;
 import co.rsk.config.BridgeRegTestConstants;
+import co.rsk.peg.bitcoin.ErpRedeemScriptBuilder;
+import co.rsk.peg.bitcoin.NonStandardErpRedeemScriptBuilderFactory;
+import co.rsk.peg.bitcoin.P2shErpRedeemScriptBuilder;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ActivationConfigsForTest;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
